Refactor: Rearrange components into a common folder for better organizatio
