+++ src/components/common/confetti-display.tsx
